+++ src/main/java/it/eng/idsa/businesslogic/routes/CamelRouteReceiver.java
@@ -19,7 +19,6 @@ import it.eng.idsa.businesslogic.processor.common.MapMultipartToIDSCP2;
 import it.eng.idsa.businesslogic.processor.common.ModifyPayloadProcessor;
 import it.eng.idsa.businesslogic.processor.common.OriginalMessageProcessor;
 import it.eng.idsa.businesslogic.processor.common.RegisterTransactionToCHProcessor;
-import it.eng.idsa.businesslogic.processor.common.SelfDescriptionProcessor;
 import it.eng.idsa.businesslogic.processor.common.TrueConnectorAuthorization;
 import it.eng.idsa.businesslogic.processor.common.ValidateTokenProcessor;
 import it.eng.idsa.businesslogic.processor.exception.ExceptionForProcessor;
@@ -113,9 +112,6 @@ public class CamelRouteReceiver extends RouteBuilder {
 	@Value("${application.dataApp.websocket.isEnabled}")
 	private boolean isEnabledDataAppWebSocket;
 	
-	@Autowired
-	private SelfDescriptionProcessor selfDescriptionProcessor;
-
 	@SuppressWarnings("unchecked")
 	@Override
 	public void configure() throws Exception {
@@ -139,7 +135,7 @@ public class CamelRouteReceiver extends RouteBuilder {
 		// Camel SSL - Endpoint: B communication http
 		if(!isEnabledWebSocket && !isEnabledIdscp2) {
 			logger.info("REST Configuration");
-			from("jetty://https4://0.0.0.0:" + configuration.getCamelReceiverPort() + "/data")
+			from("jetty://https4://0.0.0.0:" + configuration.getCamelReceiverPort() + "/data" + "?httpMethodRestrict=POST")
 				.routeId("data")
 				.process(trueConnectorAuthorization)
 				.policy("adminPolicy")
@@ -165,12 +161,6 @@ public class CamelRouteReceiver extends RouteBuilder {
                 .process(modifyPayloadProcessor)
 				.process(sendDataToBusinessLogicProcessor)
 				.removeHeaders("Camel*");
-			
-			
-			from("jetty://https4://0.0.0.0:" + configuration.getCamelReceiverPort() + "/internal/sd")
-				.routeId("internalSelfDescription")
-				.log("Requesting internal Self Description document")
-				.process(selfDescriptionProcessor);
 		} 
 		
 		if (isEnabledWebSocket) {
@@ -197,12 +187,6 @@ public class CamelRouteReceiver extends RouteBuilder {
 				.process(registerTransactionToCHProcessor)
 				.process(sendDataToBusinessLogicProcessor);
 			//@formatter:on
-			
-			
-			from("jetty://https4://0.0.0.0:" +  configuration.getWssSelfDescriptionPort() + "/internal/sd")
-				.routeId("internalSelfDescription")
-				.log("Requesting internal Self Description document")
-				.process(selfDescriptionProcessor);
 		}
 		if (isEnabledIdscp2 && receiver && !isEnabledDataAppWebSocket) {
 			logger.info("Starting IDSCP v2 Receiver route");