+++ src/main/java/it/eng/idsa/businesslogic/routes/CamelRouteSender.java
@@ -20,6 +20,7 @@ import it.eng.idsa.businesslogic.processor.common.ModifyPayloadProcessor;
 import it.eng.idsa.businesslogic.processor.common.OriginalMessageProcessor;
 import it.eng.idsa.businesslogic.processor.common.ProtocolValidationProcessor;
 import it.eng.idsa.businesslogic.processor.common.RegisterTransactionToCHProcessor;
+import it.eng.idsa.businesslogic.processor.common.SelfDescriptionProcessor;
 import it.eng.idsa.businesslogic.processor.common.ValidateTokenProcessor;
 import it.eng.idsa.businesslogic.processor.exception.ExceptionForProcessor;
 import it.eng.idsa.businesslogic.processor.exception.ExceptionProcessorSender;
@@ -131,6 +132,9 @@ public class CamelRouteSender extends RouteBuilder {
 
 	@Autowired
 	private CorrelationIDProcessor correlationIDProcessor;
+	
+	@Autowired
+	private SelfDescriptionProcessor selfDescriptionProcessor;
 
 	@SuppressWarnings("unchecked")
 	@Override
@@ -148,26 +152,31 @@ public class CamelRouteSender extends RouteBuilder {
 		onException(ExceptionForProcessor.class, RuntimeException.class)
 			.handled(true)
 			.process(processorException);
+		
+		from("jetty://https4://0.0.0.0:" + configuration.getInternalSelfdescriptionPort() + "/internal/sd" + "?httpMethodRestrict=GET")
+			.routeId("internalSelfDescription")
+			.log("Requesting internal Self Description document")
+			.process(selfDescriptionProcessor);
 
 		if(!isEnabledDataAppWebSocket) {
 			logger.info("REST self registration configuration");
-			from("jetty://https4://0.0.0.0:" + configuration.getCamelSenderPort() + "/selfRegistration/register")
+			from("jetty://https4://0.0.0.0:" + configuration.getCamelSenderPort() + "/selfRegistration/register" + "?httpMethodRestrict=POST")
 				.routeId("selfRegistration/register")
 				.process(createRegistratioMessageSender)
 				.to("direct:registrationProcess");
-			from("jetty://https4://0.0.0.0:" + configuration.getCamelSenderPort() + "/selfRegistration/update")
+			from("jetty://https4://0.0.0.0:" + configuration.getCamelSenderPort() + "/selfRegistration/update" + "?httpMethodRestrict=POST")
 				.routeId("selfRegistration/update")
 				.process(createUpdateMessageSender)
 				.to("direct:registrationProcess");
-			from("jetty://https4://0.0.0.0:" + configuration.getCamelSenderPort() + "/selfRegistration/delete")
+			from("jetty://https4://0.0.0.0:" + configuration.getCamelSenderPort() + "/selfRegistration/delete" + "?httpMethodRestrict=POST")
 				.routeId("selfRegistration/delete")
 				.process(createDeleteMessageSender)
 				.to("direct:registrationProcess");
-			from("jetty://https4://0.0.0.0:" + configuration.getCamelSenderPort() + "/selfRegistration/passivate")
+			from("jetty://https4://0.0.0.0:" + configuration.getCamelSenderPort() + "/selfRegistration/passivate" + "?httpMethodRestrict=POST")
 				.routeId("selfRegistration/passivate")
 				.process(createPassivateMessageSender)
 				.to("direct:registrationProcess");
-			from("jetty://https4://0.0.0.0:" + configuration.getCamelSenderPort() + "/selfRegistration/query")
+			from("jetty://https4://0.0.0.0:" + configuration.getCamelSenderPort() + "/selfRegistration/query" + "?httpMethodRestrict=POST")
 				.routeId("selfRegistration/query")
 				.process(createBrokerQueryMessageSender)
 				.to("direct:registrationProcess");
@@ -186,13 +195,13 @@ public class CamelRouteSender extends RouteBuilder {
 			logger.info("REST configuration");
 
 			// Camel SSL - Endpoint: A - Body binary
-			from("jetty://https4://0.0.0.0:" + configuration.getCamelSenderPort() + "/incoming-data-app/multipartMessageBodyBinary")
+			from("jetty://https4://0.0.0.0:" + configuration.getCamelSenderPort() + "/incoming-data-app/multipartMessageBodyBinary" + "?httpMethodRestrict=POST")
 				.routeId("multipartMessageBodyBinary")
 				.process(parseReceivedDataProcessorBodyBinary)
 				.to("direct:HTTP");
 					
 		    // Camel SSL - Endpoint: A - Body form-data
-		    from("jetty://https4://0.0.0.0:" + configuration.getCamelSenderPort() + "/incoming-data-app/multipartMessageBodyFormData")
+		    from("jetty://https4://0.0.0.0:" + configuration.getCamelSenderPort() + "/incoming-data-app/multipartMessageBodyFormData" + "?httpMethodRestrict=POST")
 		    	.routeId("multipartMessageBodyFormData")
 				.process(parseReceivedDataProcessorBodyFormData)
 				.to("direct:HTTP");
@@ -230,14 +239,14 @@ public class CamelRouteSender extends RouteBuilder {
 			
 			//IDSCP2 flow triggered by multipartMessageBodyBinary
 			// Camel SSL - Endpoint: A - Body binary
-			from("jetty://https4://0.0.0.0:" + configuration.getCamelSenderPort() + "/incoming-data-app/multipartMessageBodyBinary")
+			from("jetty://https4://0.0.0.0:" + configuration.getCamelSenderPort() + "/incoming-data-app/multipartMessageBodyBinary" + "?httpMethodRestrict=POST")
 				.routeId("IDSCP2 - multipartMessageBodyBinary")
 				.process(parseReceivedDataProcessorBodyBinary)
 				.to("direct:IDSCP2");
 			
 			//IDSCP2 flow triggered by multipartMessageBodyFormData
 			// Camel SSL - Endpoint: A - Body binary
-			from("jetty://https4://0.0.0.0:" + configuration.getCamelSenderPort() + "/incoming-data-app/multipartMessageBodyFormData")
+			from("jetty://https4://0.0.0.0:" + configuration.getCamelSenderPort() + "/incoming-data-app/multipartMessageBodyFormData" + "?httpMethodRestrict=POST")
 				.routeId("IDSCP2 - multipartMessageBodyFormData")
 				.process(parseReceivedDataProcessorBodyFormData)
 				.to("direct:IDSCP2");